fix check update, fix search on back press

+++ app/src/main/java/com/arny/mobilecinema/domain/interactors/update/DataUpdateInteractorImpl.kt
@@ -61,7 +61,8 @@ class DataUpdateInteractorImpl @Inject constructor(
 
     override suspend fun getUpdateDate(): Flow<DataResult<String>> = doAsync {
         var newUpdate = ""
-        if (repository.newUpdate.isBlank()) {
+        if (!repository.checkUpdate && repository.newUpdate.isBlank()) {
+            repository.checkUpdate = true
             val updateFile = withContext(Dispatchers.IO) { repository.downloadUpdate() }
             newUpdate = updateFile.readText()
             updateFile.delete()
